Political Analysis Excel Files
I used ChatGPT to generate all of this data, and then I used MySQL to create the voter_data schema and tables within it. I used ChatGPT to act as a mentor/manager in a political organization to ask me questions in order to perform the exploratory data analysis of this fictional data. I did run into some issues with the data, as the "random" generation of the data by ChatGPT allowed for a more even distribution than a realistic dataset of voters, events, and outreach.